airflow-docker-tutorial/amazon-etl/requirements.txt
apache-airflow==3.1.1
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
mysql-connector-python==9.5.0


